requirements.txt
torch==2.3.0
torchvision==0.18.0
timm==0.9.2
facenet-pytorch==2.5.3
opencv-python==4.9.0.80
numpy==1.26.4
tqdm==4.66.4
pytorch-gradcam
scikit-image==0.22.0
scikit-learn
Pillow